refactor(components): extract HelperPanelControl + HelperPanelFrame
Promotes the helper-panel chrome that the Insights view uses into the
shared @inngest/components package so the upcoming Experiments detail
view can reuse the same panel UX (collapse/expand control, framed
content area with header + close button) without duplicating it.

New shared components (ui/packages/components/src/HelperPanelControl/):
- HelperPanelControl.tsx: the rail-mounted toggle button.
- HelperPanelFrame.tsx: the panel container with icon + title + close.
- index.ts: barrel re-exports.

Dashboard call sites updated:
- InsightsHelperPanel.tsx switches to HelperPanelFrame.
- InsightsHelperPanelControl.tsx delegates to HelperPanelControl.

This is a pure extraction — Insights behavior is unchanged. Reviewers
should verify the dashboard's Insights helper panel still opens and
closes the same way.
