+++ ui/packages/components/src/HelperPanelControl/index.ts
@@ -1,0 +1,2 @@
+export { HelperPanelControl, type HelperItem } from './HelperPanelControl';
+export { HelperPanelFrame } from './HelperPanelFrame';
